Remove memory leak caused by ByteTrack
There is a memory leak in the official ByteTrack script caused by the `self.removed_stracks` variable which is allowed to grow indefinitely. This issue (&fix) was raised in the ByteTrack repo as well but the devs dont seem to be active anymore and arent reviewing pull requests. The Ultralytics yolov8 repo also had this same bug and together with their devs, we fixed it by adding these two lines which simply keep a fixed length of 100 to the history of removed tracks. This effectively fixes the memory leak without any impact on the accuracy of the tracker.